fix broken mcm tests and universalize package management
