+++ h2integrate/converters/co2/marine/test/test_doc.py
@@ -3,14 +3,35 @@ import pytest
 import openmdao.api as om
 from openmdao.utils.assert_utils import assert_near_equal
 
+from conftest import has_mcm
 from h2integrate.converters.co2.marine.direct_ocean_capture import DOCPerformanceModel
 
 
-has_mcm = True
-try:
-    import mcm  # noqa: F401
-except ModuleNotFoundError:
-    has_mcm = False
+@pytest.fixture
+def tech_config():
+    return {
+        "model_inputs": {
+            "performance_parameters": {
+                "power_single_ed_w": 24000000.0,  # W
+                "flow_rate_single_ed_m3s": 0.6,  # m^3/s
+                "number_ed_min": 1,
+                "number_ed_max": 10,
+                "E_HCl": 0.05,  # kWh/mol
+                "E_NaOH": 0.05,  # kWh/mol
+                "y_ext": 0.9,
+                "y_pur": 0.2,
+                "y_vac": 0.6,
+                "frac_ed_flow": 0.01,
+                "use_storage_tanks": True,
+                "initial_tank_volume_m3": 0.0,  # m^3
+                "store_hours": 12.0,  # hours
+                "sal": 33.0,  # ppt
+                "temp_C": 12.0,  # degrees Celsius
+                "dic_i": 0.0022,  # mol/L
+                "pH_i": 8.1,  # initial pH
+            },
+        },
+    }
 
 
 @pytest.mark.unit
